+++ gnukek_cli/command_handlers/export.py
@@ -6,9 +6,9 @@ from dependency_injector.wiring import Provide, inject
 
 from gnukek_cli.config import SettingsProvider
 from gnukek_cli.container import Container
-from gnukek_cli.exceptions import KeyNotFoundError
 from gnukek_cli.keys.storages import PrivateKeyStorage, PublicKeyStorage
-from gnukek_cli.passwords import PasswordPrompt
+from gnukek_cli.utils.exceptions import KeyNotFoundError
+from gnukek_cli.utils.passwords import PasswordPrompt
 
 logger = logging.getLogger(__name__)
