+++ gnukek_cli/utils/exceptions.py
