+++ tests/test_command_handlers/test_sign.py
@@ -4,7 +4,7 @@ from io import BytesIO
 import pytest
 
 from gnukek_cli.command_handlers.sign import SignContext, SignHandler
-from gnukek_cli.exceptions import KeyNotFoundError
+from gnukek_cli.utils.exceptions import KeyNotFoundError
 from tests.constants import KEY_ID, SAMPLE_MESSAGE
